Respond to review commemt.

+++ runtime/ints.c
@@ -675,7 +675,7 @@ CAMLprim int64_t caml_int64_of_string_unboxed(value s)
       if (res >  (uint64_t)1 << 63) caml_failwith(INT64_ERRMSG);
     }
   }
-  if (sign < 0) res = - res;
+  if (sign < 0) res = -(int64_t)res;
   return res;
 }
